multi-agent-dev-team/requirements.txt
python-dotenv
openai>=1.68.2
requests
openai-agents